comctl32/edit: Get rid of useless casts.

+++ dlls/comctl32/edit.c
@@ -5017,7 +5017,7 @@ static LRESULT CALLBACK EDIT_WindowProc(HWND hwnd, UINT msg, WPARAM wParam, LPAR
     case WM_MOUSEWHEEL:
     {
         int wheelDelta;
-        UINT pulScrollLines = 3;
+        INT pulScrollLines = 3;
         SystemParametersInfoW(SPI_GETWHEELSCROLLLINES,0, &pulScrollLines, 0);
 
         if (wParam & (MK_SHIFT | MK_CONTROL))
@@ -5037,9 +5037,9 @@ static LRESULT CALLBACK EDIT_WindowProc(HWND hwnd, UINT msg, WPARAM wParam, LPAR
         if (es->wheelDeltaRemainder && pulScrollLines)
         {
             int cLineScroll;
-            pulScrollLines = (int) min((UINT) es->line_count, pulScrollLines);
-            cLineScroll = pulScrollLines * (float)es->wheelDeltaRemainder / WHEEL_DELTA;
-            es->wheelDeltaRemainder -= WHEEL_DELTA * cLineScroll / (int)pulScrollLines;
+            pulScrollLines = min(es->line_count, pulScrollLines);
+            cLineScroll = pulScrollLines * es->wheelDeltaRemainder / WHEEL_DELTA;
+            es->wheelDeltaRemainder -= WHEEL_DELTA * cLineScroll / pulScrollLines;
             result = EDIT_EM_LineScroll(es, 0, -cLineScroll);
         }
         break;
